Implemented different food sources to /eat from
